+++ src/tests/PerformanceTests/Memory/SafeString/SafeStringPerformanceTests.cs
@@ -20,7 +20,7 @@ namespace SafeOrbit.Memory
             // Act
             var actualPerformance = Measure(() =>
             {
-                _ = sut.ToSafeBytesAsync();
+                _ = sut.GetHashCode();
             }, 10);
 
             // Assert
@@ -29,19 +29,19 @@ namespace SafeOrbit.Memory
 
 
         [Test]
-        public async Task ToSafeBytes_1000Chars_TakesLessThan100ms()
+        public async Task RevealDecryptedBytesAsync_1000Chars_TakesLessThan5000ms()
         {
             // Arrange
-            const int expectedHigherLimit = 100;
+            const int expectedHigherLimit = 5000;
             SafeOrbitCore.Current.StartEarly();
             var sut = GetSut();
             await sut.AppendAsync(new string('m', 1000));
 
             // Act
-            var actualPerformance = Measure(() =>
+            var actualPerformance = await MeasureAsync(async () =>
             {
-                _ = sut.GetHashCode();
-            }, 10);
+                _ = await sut.RevealDecryptedBytesAsync();
+            }, 2);
 
             // Assert
             Assert.That(actualPerformance, Is.LessThanOrEqualTo(expectedHigherLimit));